Changes to allow libdragon to link with Rust
* Change tools to use mips64-libdragon-elf
* Change libdragon to use n32 ABI
* Build libdragon with -mno-gpopt
* Convert dashes to underscores in rsp build rule

LLVM doesn't respect $gp and Rust won't build with -mno-abicalls for now. So gpopt is disabled
in libdragon and newlib for now.

* Update build-toolchain-matrix.yml

Disable .deb and .rpm builds and introduce raw .zip build

* Increase the size of joypad_buttons_t

This is required to convince gcc to return the struct in $a0, rather
than $v0/$v1

* Disable .eh_frame and gcc exception frame handling

This shouldn't be necessary but it is

* Need wrappers for structs passed as value

LLVM doesn't use 64-bit registers when passing structs by value to
function calls

* Changing tabs to spaces in some doxygen comments

doxygen-rs panics on this input. I think it expects at least one space

* Keep inline functions when producing the library